--- app/views/notes/_comment.html.erb
@@ -1,3 +1,0 @@
-<% @comments.each do |comment| %>
-  <p><%= comment.content %></p>
-<% end %>
