--- db/migrate/20210525015934_create_comments.rb
@@ -1,10 +1,0 @@
-class CreateComments < ActiveRecord::Migration[6.0]
-  def change
-    create_table :comments do |t|
-      t.string     :content
-      t.references :note,           foreign_key: true
-      t.references :item,           foreign_key: true
-      t.timestamps
-    end
-  end
-end
