+++ spec/jobs/message_broadcast_job_spec.rb
@@ -1,5 +1,5 @@
 require 'rails_helper'
 
-RSpec.describe Comment, type: :model do
+RSpec.describe MessageBroadcastJob, type: :job do
   pending "add some examples to (or delete) #{__FILE__}"
 end
